feat: replaces old physics with improved physics for multi collisions
Old Physics could only detect when the ball hit one brick. However
ther ball can hit several game objects at once. New Physics implementation
takes that into account by checking all the game objects for collisions
with the ball before the reflection is handled
